app/requirements.txt
Django
django-cors-headers
django-redis
djangorestframework
djangorestframework-jwt
python-dotenv
psycopg2-binary
gunicorn





